refactor: use super:: instead of crate::internal
