+++ src/internal/parser.rs
@@ -1,7 +1,7 @@
-use crate::internal::error_reporter::ErrorReporter;
-use crate::internal::expr::Expr;
-use crate::internal::token::{Literal, Token};
-use crate::internal::token_type;
+use super::error_reporter::ErrorReporter;
+use super::expr::Expr;
+use super::token::{Literal, Token};
+use super::token_type;
 
 #[derive(Debug, Clone)]
 pub struct ParseError;
